+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -104,7 +104,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01ke9dbnv1fzr4je28nx4zw05k"
+        "conv_01ke9hjveacqwbnp37qserzewr"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """