+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1166,7 +1166,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-06T10:24:07.014Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-06T11:37:56.383Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1218,7 +1218,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-06T10:24:07.014Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-06T11:37:56.382Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1270,7 +1270,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[Source] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-06T10:24:07.013Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-06T11:37:56.382Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1349,5 +1349,5 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-06T10:23:54.828Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-06T11:37:44.678Z")
     r"""The date and time the resource was last updated"""