+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -239,7 +239,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KE9DBP8010MRDBACZ002ZZ7Q"
+        "01KE9HJVST2K484KTQB2C2ANA0"
     )
     r"""The unique identifier of the data source"""