+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -130,7 +130,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KE9DBP7X183F424JY31R03M7"
+        "01KE9HJVSQGJZS7MC03C3DMZC9"
     )
     r"""The unique identifier of the data source"""