+++ packages/orq-rc/src/orq_ai_sdk/models/retrieveconversationop.py
@@ -112,7 +112,7 @@ class RetrieveConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01ke9dbnv3wabypnsgws265h8w"
+        "conv_01ke9hjvedx1mmc7v4rf73bm54"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """