+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KE9DBP7ZVKEFFEFMXE0DYKBX"
+        "01KE9HJVSSHPVRG07QVA4NZ4FW"
     )
     r"""The unique identifier of the data source"""