+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -150,7 +150,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01ke9dbnv55wq7hs42pz98z1n1"
+        "conv_01ke9hjveeqkmy4hb9dvt879vc"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """