+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KE9DBP82NEV6B0GGAVA844B3"
+        "01KE9HJVT226165FN3G77KFT8Q"
     )
     r"""The unique identifier of the data source"""