+++ packages/orq-rc/RELEASES.md
@@ -4559,3 +4559,13 @@ Based on:
 - [python v4.2.0-rc.11] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.11] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.11 - packages/orq-rc
+
+## 2026-01-06 11:40:11
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.681.0 (2.789.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.12] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.12] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.12 - packages/orq-rc